perform tab repeats before soloPanelContext
